refactor(backend): normalize Yacht into backend/yacht/ package (#154)
Moves Yacht game code from flat files into a subpackage matching the
structure of the other four games (blackjack, cascade, ludo, twenty48).

Changes:
- backend/game.py            → backend/yacht/game.py
- backend/models.py          → backend/yacht/models.py
- Yacht routes + session store extracted from main.py into
  backend/yacht/router.py (mirrors blackjack/ludo/twenty48 router
  structure)
- main.py now just imports and mounts the yacht router with
  `prefix="/game"` — all paths and behavior preserved
- reset_game() test helper exposed from yacht/router.py (matching
  blackjack's pattern); tests no longer reach into main._sessions

Behavior preserved — all /game/* endpoint paths, rate limits, session
handling, and state response shapes are identical. Renaming the prefix
from /game/* to /yacht/* is tracked separately in #161 (breaking API
change).

399/399 backend tests pass, 97% coverage, black clean.
