+++ backend/main.py
@@ -2,9 +2,8 @@ import json
 import logging
 import os
 import time
-from collections import OrderedDict
 
-from fastapi import FastAPI, HTTPException, Request, Response
+from fastapi import FastAPI, Request, Response
 from fastapi.middleware.cors import CORSMiddleware
 from fastapi.responses import JSONResponse
 from slowapi.errors import RateLimitExceeded
@@ -15,14 +14,12 @@ import sentry_sdk
 from sentry_sdk.integrations.fastapi import FastApiIntegration
 from sentry_sdk.integrations.starlette import StarletteIntegration
 
-from game import YachtGame
 from limiter import _real_ip, limiter
-from models import GameStateResponse, PossibleScoresResponse, RollRequest, ScoreRequest
-from session import get_session_id
 from cascade.router import router as cascade_router
 from blackjack.router import router as blackjack_router
 from ludo.router import router as ludo_router
 from twenty48.router import router as twenty48_router
+from yacht.router import router as yacht_router
 
 # ---------------------------------------------------------------------------
 # Audit logger — emits JSON lines; Render's log aggregator handles timestamps
@@ -48,6 +45,8 @@ if _sentry_dsn:
 # ---------------------------------------------------------------------------
 
 app = FastAPI(title="Gaming App API")
+# Yacht still uses /game/* prefix for backwards compat — rename tracked in #161.
+app.include_router(yacht_router, prefix="/game")
 app.include_router(cascade_router, prefix="/cascade")
 app.include_router(blackjack_router, prefix="/blackjack")
 app.include_router(ludo_router, prefix="/ludo")
@@ -167,99 +166,6 @@ async def security_headers(request: Request, call_next) -> Response:
     if "server" in response.headers:
         del response.headers["server"]
     return response
-
-
-# ---------------------------------------------------------------------------
-# Session-isolated Yacht state
-# ---------------------------------------------------------------------------
-
-_MAX_SESSIONS = 500
-_sessions: OrderedDict[str, YachtGame | None] = OrderedDict()
-
-
-def _evict_if_full() -> None:
-    while len(_sessions) >= _MAX_SESSIONS:
-        _sessions.popitem(last=False)
-
-
-def _get_game(session_id: str) -> YachtGame:
-    game = _sessions.get(session_id)
-    if game is None:
-        raise HTTPException(status_code=404, detail="No game in progress. POST /game/new first.")
-    return game
-
-
-def _state_response(game: YachtGame) -> GameStateResponse:
-    return GameStateResponse(
-        dice=game.dice,
-        held=game.held,
-        rolls_used=game.rolls_used,
-        round=game.round,
-        scores=game.scores,
-        game_over=game.game_over,
-        upper_subtotal=game.upper_subtotal(),
-        upper_bonus=game.upper_bonus(),
-        yacht_bonus_count=game.yacht_bonus_count,
-        yacht_bonus_total=game.yacht_bonus_total(),
-        total_score=game.total_score(),
-    )
-
-
-# ---------------------------------------------------------------------------
-# Routes
-# ---------------------------------------------------------------------------
-
-
-@app.post("/game/new", response_model=GameStateResponse)
-@limiter.limit("10/minute")
-def new_game(request: Request) -> GameStateResponse:
-    sid = get_session_id(request)
-    _evict_if_full()
-    _sessions[sid] = YachtGame()
-    return _state_response(_sessions[sid])
-
-
-@app.get("/game/state", response_model=GameStateResponse)
-@limiter.limit("60/minute")
-def get_state(request: Request) -> GameStateResponse:
-    sid = get_session_id(request)
-    return _state_response(_get_game(sid))
-
-
-@app.post("/game/roll", response_model=GameStateResponse)
-@limiter.limit("30/minute")
-def roll(request: Request, body: RollRequest) -> GameStateResponse:
-    sid = get_session_id(request)
-    game = _get_game(sid)
-    if len(body.held) != 5:
-        raise HTTPException(status_code=422, detail="'held' must have exactly 5 booleans.")
-    try:
-        game.roll(body.held)
-    except ValueError as e:
-        raise HTTPException(status_code=400, detail=str(e))
-    return _state_response(game)
-
-
-@app.post("/game/score", response_model=GameStateResponse)
-@limiter.limit("20/minute")
-def score(request: Request, body: ScoreRequest) -> GameStateResponse:
-    sid = get_session_id(request)
-    game = _get_game(sid)
-    try:
-        game.score(body.category)
-    except ValueError as e:
-        raise HTTPException(status_code=400, detail=str(e))
-    return _state_response(game)
-
-
-@app.get("/game/possible-scores", response_model=PossibleScoresResponse)
-@limiter.limit("60/minute")
-def possible_scores(request: Request) -> PossibleScoresResponse:
-    sid = get_session_id(request)
-    game = _get_game(sid)
-    if game.rolls_used == 0:
-        return PossibleScoresResponse(possible_scores={})
-    return PossibleScoresResponse(possible_scores=game.possible_scores())
 
 
 @app.get("/health")
